LG-7819 Load test USPS getProofingResults endpoint
**Why**:
-This forms the first stub of a series of tests that is intended to test
USPS getProofingResults api endpoint. This particular test is applicable
to the identity-fake-server application and will need additional work and
setup between the test IPP database and USPS api endpoint.

**How**:
-The usps_get_proofing_result.locustfile.py includes two environment
variables. NUM_USERS and ENV. ENV is the environment targeted (for a local
running identity-fake-server, it is http://0.0.0.0:5555). The constant
ENDPOINT is the USPS endpoint for getProofingResults.

changelog: Internal, Loadtesting, New test for integration to USPS

On branch micah/lg-7819-usps-get-proofing-result-test
Changes to be committed:
        new file:   load_testing/usps_get_proofing_result.locustfile.py